Speed up Q43
Since the join operation is so costly add another sanity check before it.
The triple d4,d5,d6 must be divisible by five so its final digit
must be a 0 or 5 and can be quickly tested.

Also reverse the order of the prime tests. This is a slight speed up but
allows elimination on the larger primes first. For numbers 1 - 999 only
58 are divisible by 17.